refactor(tui): extract bounded terminal fixture support
