+++ src/tui/tui-pty-harness.e2e.test.ts
@@ -1,15 +1,17 @@
 // Exercises the fake-backend TUI PTY harness and visible terminal output.
-import { mkdtemp, readFile, rm, writeFile } from "node:fs/promises";
+import { mkdtemp, rm, writeFile } from "node:fs/promises";
 import { tmpdir } from "node:os";
 import path from "node:path";
 import { pathToFileURL } from "node:url";
 import { afterAll, beforeAll, describe, expect, it } from "vitest";
+import {
+  buildOpaqueSessionIsolationFixture,
+  objectFieldEquals,
+  readFixtureLog,
+  waitForFixtureLogEntry,
+  type FixtureLogEntry,
+} from "./tui-pty-harness-fixture-support.js";
 import { sleep, startPty, type PtyRun } from "./tui-pty-test-support.js";
-
-type FixtureLogEntry = {
-  method: string;
-  payload?: unknown;
-};
 
 const activeRuns: PtyRun[] = [];
 const STARTUP_TIMEOUT_MS = 20_000;
@@ -17,53 +19,6 @@ const OUTPUT_TIMEOUT_MS = 2_000;
 const EXIT_TIMEOUT_MS = 4_000;
 const TEST_TIMEOUT_MS = 5_000;
 const STARTUP_TEST_TIMEOUT_MS = 25_000;
-
-async function readFixtureLog(logPath: string): Promise<FixtureLogEntry[]> {
-  try {
-    const text = await readFile(logPath, "utf8");
-    return text
-      .split("\n")
-      .filter(Boolean)
-      .map((line) => JSON.parse(line) as FixtureLogEntry);
-  } catch (error) {
-    if ((error as NodeJS.ErrnoException).code === "ENOENT") {
-      return [];
-    }
-    throw error;
-  }
-}
-
-async function waitForFixtureLogEntry(
-  logPath: string,
-  predicate: (entry: FixtureLogEntry) => boolean,
-  timeoutMs = OUTPUT_TIMEOUT_MS,
-  readPtyOutput?: () => string,
-) {
-  const start = Date.now();
-  while (Date.now() - start < timeoutMs) {
-    const entries = await readFixtureLog(logPath);
-    const match = entries.find(predicate);
-    if (match) {
-      return match;
-    }
-    await sleep(25);
-  }
-  const entries = await readFixtureLog(logPath);
-  // A swallowed command leaves no RPC behind, so the RPC log alone cannot say
-  // whether the TUI rejected the input; the terminal output carries that reason.
-  const ptyOutput = readPtyOutput?.() ?? "";
-  throw new Error(
-    `timed out waiting for fixture log entry\n${JSON.stringify(entries, null, 2)}\n${ptyOutput}`,
-  );
-}
-
-function objectFieldEquals(entry: FixtureLogEntry, field: string, value: unknown) {
-  if (typeof entry.payload !== "object" || entry.payload === null) {
-    return false;
-  }
-  const payload = entry.payload as Record<string, unknown>;
-  return Object.hasOwn(payload, field) && payload[field] === value;
-}
 
 async function writeTuiPtyFixtureScript(dir: string) {
   // Temp files sit outside the repo package scope; .mts preserves the ESM contract under tsx.
@@ -237,36 +192,7 @@ async function writeTuiPtyFixtureScript(dir: string) {
             });
             return { runId };
           }
-          if (opts.message.startsWith("opaque session isolation proof: ")) {
-            const otherSessionKey = opts.sessionKey.includes(":matrix:")
-              ? opts.sessionKey.replace("!MixedRoomAbCdEf", "!mixedroomabcdef")
-              : opts.sessionKey.replace("AbC123=", "abc123=");
-            const marker = "PTY_FOREIGN_OPAQUE_SESSION_MESSAGE";
-            queueMicrotask(() => {
-              record("foreignSessionEvent", { sessionKey: otherSessionKey, marker });
-              this.onEvent?.({
-                event: "chat",
-                payload: {
-                  runId: "run-foreign-opaque-session",
-                  sessionKey: otherSessionKey,
-                  state: "delta",
-                  message: {
-                    role: "assistant",
-                    content: [{ type: "text", text: marker }],
-                  },
-                },
-              });
-              this.onEvent?.({
-                event: "session.message",
-                payload: {
-                  agentId: "main",
-                  sessionKey: otherSessionKey,
-                  sessionId: "foreign-opaque-session",
-                  updatedAt: Date.now(),
-                },
-              });
-            });
-          }
+          ${buildOpaqueSessionIsolationFixture()}
           const responseDelayMs =
             opts.message === "slow prompt" ||
             opts.message === "slow reset proof" ||
@@ -562,7 +488,7 @@ async function startTuiFixture(opts: { env?: NodeJS.ProcessEnv } = {}) {
     run,
     logPath,
     waitForLogEntry: async (predicate: (entry: FixtureLogEntry) => boolean, timeoutMs?: number) =>
-      await waitForFixtureLogEntry(logPath, predicate, timeoutMs, run.output),
+      await waitForFixtureLogEntry(logPath, predicate, timeoutMs ?? OUTPUT_TIMEOUT_MS, run.output),
     cleanup: async () => {
       await run.dispose();
       await rm(tempDir, { recursive: true, force: true });
